updated resourses

+++ src/AdminResources/Contacts.php
@@ -3,6 +3,7 @@
 namespace Daveismyname\MsGraph\AdminResources;
 
 use Daveismyname\MsGraph\Facades\MsGraphAdmin;
+use Exception;
 
 class Contacts extends MsGraphAdmin
 {
@@ -30,6 +31,10 @@ class Contacts extends MsGraphAdmin
 
 	public function get($params = [])
 	{
+        if ($this->userId == null) {
+            throw new Exception("userId is required.");
+        }
+
         if ($params == []) {
 
             $top = request('top', $this->top);
@@ -59,21 +64,37 @@ class Contacts extends MsGraphAdmin
 
     public function find($id)
     {
+        if ($this->userId == null) {
+            throw new Exception("userId is required.");
+        }
+
         return MsGraphAdmin::get('users/'.$this->userId.'/contacts/'.$id);
     }
 
     public function store(array $data)
     {
+        if ($this->userId == null) {
+            throw new Exception("userId is required.");
+        }
+
         return MsGraphAdmin::post('users/'.$this->userId.'/contacts', $data);
     }
 
     public function update($id, array $data)
     {
+        if ($this->userId == null) {
+            throw new Exception("userId is required.");
+        }
+
         return MsGraphAdmin::patch('users/'.$this->userId.'/contacts/'.$id, $data);
     }
 
     public function delete($id)
     {
+        if ($this->userId == null) {
+            throw new Exception("userId is required.");
+        }
+
         return MsGraphAdmin::delete('users/'.$this->userId.'/contacts/'.$id);
     }
 }
